+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Add `getSqlBackendPool`
 * Move `rerunnableLift` into `MonadRerunnableTrans` typeclass in `Control.Monad.Trans.Rerunnable`
 * Add `catchSqlTransaction`
+* Add `retryCallback` to `SqlQueryEnv`
 
 # 0.4.0.0